#1298 - Dynamic instrumentation with javaagent is broken - tests
