+++ activejdbc-instrumentation-test/pom.xml
@@ -59,8 +59,7 @@
                         <activejdbc.password>${jdbc.password}</activejdbc.password>
                         <activejdbc.driver>${jdbc.driver}</activejdbc.driver>
                     </systemPropertyVariables>
-                    <forkMode>pertest</forkMode>
-                    <argLine>-javaagent:${pom.basedir}/target/activejdbc-instrumentation-3.1-SNAPSHOT.jar</argLine>
+                    <argLine>-javaagent:${basedir}/target/activejdbc-instrumentation-3.1-SNAPSHOT.jar</argLine>
                     <workingDirectory>${basedir}/target</workingDirectory>
                 </configuration>
             </plugin>
