+++ activejdbc-instrumentation-test/src/main/java/org/javalite/activejdbc/instrumentation/tests/models/Book.java
@@ -15,11 +15,15 @@ limitations under the License.
 */
 
 
-package org.javalite.activejdbc.models;
+package org.javalite.activejdbc.instrumentation.tests.models;
 
 import org.javalite.activejdbc.Model;
+import org.javalite.activejdbc.annotations.BelongsTo;
 
-/**
- * @author Igor Polevoy
- */
-public class Reader extends Model {}
+
+@BelongsTo(parent = Library.class, foreignKeyName = "lib_id")
+public class Book  extends Model {
+    static {
+        validatePresenceOf("title", "author");
+    }
+}
