--- activejdbc-instrumentation-test/src/test/projects/javaagent/pom.xml
@@ -1,90 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>org.javalite</groupId>
-    <artifactId>javaagent-test</artifactId>
-    <version>3.1-SNAPSHOT</version>
-
-    <name>JavaLite - ActiveJDBC JavaAgent Instrumentation Test</name>
-
-    <properties>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <jdbc.driver>org.h2.Driver</jdbc.driver>
-        <jdbc.url>jdbc:h2:${project.basedir}/target/test-db</jdbc.url>
-        <jdbc.user>sa</jdbc.user>
-        <jdbc.password />
-        <db>h2</db>
-        <h2.version>2.1.210</h2.version>
-    </properties>
-
-    <build>
-        <plugins>
-            <plugin>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.1</version>
-                <configuration>
-                    <source>16</source>
-                    <target>16</target>
-                    <!--
-                        Alternate config, working on Java11
-                        <release>11</release>
-                    -->
-                    <meminitial>128m</meminitial>
-                    <maxmem>512m</maxmem>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-enforcer-plugin</artifactId>
-                <version>3.0.0-M2</version>
-                <executions>
-                    <execution>
-                        <id>enforce-versions</id>
-                        <goals>
-                            <goal>enforce</goal>
-                        </goals>
-                        <configuration>
-                            <rules>
-                                <requireJavaVersion>
-                                    <version>1.8</version>
-                                </requireJavaVersion>
-                            </rules>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-source-plugin</artifactId>
-            </plugin>
-            <plugin>
-                <groupId>org.codehaus.mojo</groupId>
-                <artifactId>exec-maven-plugin</artifactId>
-                <configuration>
-                    <executable>java -javaagent:/home/yan/projects/javalite/activejdbc-instrumentation/target/activejdbc-instrumentation-3.1-SNAPSHOT.jar</executable>
-                    <mainClass>org.javalite.activejdbc.Main</mainClass>
-                    <arguments>
-                        <argument>-javaagent:/home/yan/projects/javalite/activejdbc-instrumentation/target/activejdbc-instrumentation-3.1-SNAPSHOT.jar</argument>
-                    </arguments>
-                </configuration>
-
-            </plugin>
-        </plugins>
-    </build>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.javalite</groupId>
-            <artifactId>activejdbc</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.h2database</groupId>
-            <artifactId>h2</artifactId>
-            <version>${h2.version}</version>
-        </dependency>
-    </dependencies>
-
-</project>
