--- activejdbc-instrumentation-test/src/test/projects/javaagent/src/main/java/org/javalite/activejdbc/models/Book.java
@@ -1,29 +1,0 @@
-/*
-Copyright 2009-2019 Igor Polevoy
-
-Licensed under the Apache License, Version 2.0 (the "License"); 
-you may not use this file except in compliance with the License. 
-You may obtain a copy of the License at 
-
-http://www.apache.org/licenses/LICENSE-2.0 
-
-Unless required by applicable law or agreed to in writing, software 
-distributed under the License is distributed on an "AS IS" BASIS, 
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
-See the License for the specific language governing permissions and 
-limitations under the License. 
-*/
-
-
-package org.javalite.activejdbc.models;
-
-import org.javalite.activejdbc.Model;
-import org.javalite.activejdbc.annotations.BelongsTo;
-
-
-@BelongsTo(parent = Library.class, foreignKeyName = "lib_id")
-public class Book  extends Model {
-    static {
-        validatePresenceOf("title", "author");
-    }
-}
