--- activejdbc-instrumentation-test/src/test/projects/javaagent/src/main/resources/database.properties
@@ -1,12 +1,0 @@
-
-development.driver=${jdbc.driver}
-development.username=${jdbc.user}
-development.password=${jdbc.password}
-development.url=${jdbc.url}
-
-
-#this one is used by DBSpecTest
-development.test.driver=${jdbc.driver}
-development.test.username=${jdbc.user}
-development.test.password=${jdbc.password}
-development.test.url=${jdbc.url}
